requirements.txt
python-telegram-bot[job-queue]>=21.0.1
groq>=0.9.0
supabase>=2.11.0
fpdf2>=2.7.9
httpx
python-dotenv
PyMuPDF
python-jobspy>=1.1.80
beautifulsoup4>=4.12.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
